upgrade notice fixed v2
